Add support for default styling (#9)
* Allow default styling

Create a new style property at config level. This property allows to configure a default colour for the graph when: (1) there is no severity defined, or (2) the severities do not match the current value.

* Update README to document default styling

* Update README default value for style property

* Add documentation for font colour

This branch adds support for adding colour to both default and severity changes.

* Update card to support font colour changes

Colour changes are supported on both severity and default level.

* Remove debugging logs